refactor: layout signin company
