+++ src/app/candidato/(auth)/login/page.jsx
@@ -1,4 +1,3 @@
-'use client';
 import Link from 'next/link';
 import { Description } from './Description';
 import { Logo } from './Logo';
@@ -16,7 +15,7 @@ const SignIn = () => {
         <Link href="/" className="mb-40 cursor-pointer">
           <Logo />
         </Link>
-        <div className="flex flex-col w-4/5 gap-4">
+        <div className="flex flex-col w-5/6 gap-4">
           <Title variant="inverse">{candidate.signin.title}</Title>
           <Description>{candidate.signin.description}</Description>
         </div>
@@ -25,7 +24,7 @@ const SignIn = () => {
       <Layout.Right className="flex flex-col items-center" variant="inverse">
         <div className="lg:pl-10  w-full h-full flex flex-col items-center">
           <Navbar />
-          <div className="w-full grow flex flex-col items-center justify-center  gap-6 lg:gap-8">
+          <div className="w-full grow flex flex-col items-center justify-center gap-6 lg:gap-8">
             <Title className="text-4xl lg:text-5xl" variant="inverseForm">
               {candidate.signin.form.title}
             </Title>
