+++ src/app/empresa/(auth)/login/Logo.jsx
@@ -25,4 +25,5 @@ const Logo = () => {
     />
   );
 };
+
 export { Logo };
