+++ src/app/empresa/(auth)/login/page.jsx
@@ -1,84 +1,44 @@
-'use client';
-import Image from 'next/image';
 import Link from 'next/link';
-import { useRouter } from 'next/navigation';
-import { twMerge } from 'tailwind-merge';
+import { Description } from './Description';
+import { Logo } from './Logo';
+import { Navbar } from './Navbar';
 import { SigninForm } from '@/components/company/Signin/SigninForm';
-import { AuthNavbar } from '@/components/shared/AuthNavbar';
 import { ButtonLink } from '@/components/shared/ButtonLink';
 import { Layout } from '@/components/shared/Layout';
 import { Title } from '@/components/shared/Title';
-import { themes, useTheme } from '@/contexts/ThemeContext';
 import { company } from '@/locales';
 
-const styles = {
-  default: {
-    logo: {
-      [themes.DEFAULT]: 'logo_recrutaeu_purple',
-      [themes.DARK]: 'logo_recrutaeu_black',
-      [themes.LIGHT]: 'logo_recrutaeu_white',
-    },
-    text: {
-      [themes.DEFAULT]: 'text-primary-90',
-      [themes.DARK]: 'text-neutral-90',
-      [themes.LIGHT]: 'text-neutral-0',
-    },
-    description: {
-      [themes.DEFAULT]: 'text-neutral-0',
-      [themes.DARK]: 'text-neutral-0',
-      [themes.LIGHT]: 'text-neutral-90',
-    },
-    titleForm: {
-      [themes.DEFAULT]: 'text-primary-40',
-      [themes.DARK]: 'text-neutral-0',
-      [themes.LIGHT]: 'text-neutral-90',
-    },
-  },
-};
-
-const SignIn = ({ variant = 'default' }) => {
-  const { theme } = useTheme();
-  const style = styles[variant];
-  const route = useRouter();
-
+const SignIn = () => {
   return (
     <>
       <Layout.Left className="hidden lg:flex flex-col" variant="inverse">
         <Link href="/" className="mb-40 cursor-pointer">
-          <Image
-            src={`/assets/images/${style.logo[theme]}.png`}
-            width={120}
-            height={120}
-            alt="logo recrutaeu"
-          />
+          <Logo />
         </Link>
-        <div className="flex flex-col  gap-4">
+        <div className="flex flex-col w-5/6 gap-4">
           <Title variant="inverse">{company.signin.title}</Title>
-          <p className={twMerge('text-xl', style.text[theme])}>{company.signin.description}</p>
+          <Description className="text-xl">{company.signin.description}</Description>
         </div>
       </Layout.Left>
 
       <Layout.Right className="flex flex-col items-center" variant="inverse">
-        <AuthNavbar
-          variant="inverse"
-          onBack={() => {
-            route.push('/');
-          }}
-        />
-        <div className="w-full grow flex flex-col items-center justify-center gap-8">
-          <Title className={twMerge('text-5xl', style.titleForm[theme])}>
-            {company.signin.form.title}
-          </Title>
+        <div className="lg:pl-10  w-full h-full flex flex-col items-center">
+          <Navbar />
+          <div className="w-full grow flex flex-col items-center justify-center gap-6  lg:gap-8">
+            <Title className="text-4xl lg:text-5xl" variant="inverseForm">
+              {company.signin.form.title}
+            </Title>
 
-          <SigninForm variant="inverse" />
-        </div>
-        <div className="w-full flex justify-center gap-2">
-          <p className={twMerge('text-lg', style.description[theme])}>
-            {company.signin.buttonLink.description}
-          </p>
-          <ButtonLink href="/empresa/cadastro" variant="inverse" className="text-lg">
-            {company.signin.buttonLink.label}
-          </ButtonLink>
+            <SigninForm variant="inverse" />
+          </div>
+          <div className="w-full flex justify-center gap-2">
+            <Description className="text-lg" variant="inverse">
+              {company.signin.buttonLink.description}
+            </Description>
+            <ButtonLink href="/empresa/cadastro" variant="inverse" className="text-lg">
+              {company.signin.buttonLink.label}
+            </ButtonLink>
+          </div>
         </div>
       </Layout.Right>
     </>
