+++ src/components/company/Signin/SigninForm.jsx
@@ -1,3 +1,4 @@
+'use client';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
